info: improve AppArmor detection using runc's helper
Currently, CRI-O uses the IsSupported() method to determine
whether AppArmor is enabled. However, IsSupported() checks
more than just the existence of the AppArmor parser — it also
considers whether the environment is rootless and whether
runc is enabled. This leads to incorrect reporting in
scenarios where AppArmor is enabled on the host but the
environment is rootless, causing AppArmor to be shown as
disabled.

This patch updates the Info struct to utilize runc's isEnabled() helper
for determining whether AppArmor is enabled. Unlike previous checks,
this method directly reads from /sys/module/apparmor/parameters/enabled,
providing a more reliable indication of the AppArmor status on the host.

Signed-off-by: Feruzjon Muyassarov <feruzjon.muyassarov@est.tech>